Update README.md
Added a new section at the bottom of this README.md document, for the Update from November 26th, 2023. I talk about my process of Documentation some more, how I work in general in the Art Operation at  Historiotheque. I mention how I post on social media, but in a differentiated manner, i.e. not the same content on every platform, platform-specific exclusive content, catered to the platform. I also mentioned that the Historiotheque itself gets BUILT UP, THEN TORN DOWN, similar to my Work Sessions in The Art Operation at any given moment in time, in the EVER-CHANGING WORKSPACE. I end with an image of the "OPEN THE BOX" SCHEMA FOR THE ART-OPS-IN-A-BOX GITHUB REPOSITORY.