Create README.md
Starting a new Folder in the GeneralWorkflow Folder in the METHODS Folder in the Documentation GitHub Repository. Writing about the SWITCHBOARD METHOD where you take n number of tasks and work on a given task at a given time, and move from task to task in a pseudo-random temporal order or sequence, in an essentially non-linear manner of work organization.