Update README.md
Same as before. Corrected a clerical error. I had used the term "ready-and-near-at-hand" when I meant to use the Heideggerian terminology of just "ready-to-hand". So I corrected it. It's important. Error introduction can be dangerous. The entire history of the world could go off-track with just a 1-bit error which I call an n-bit or noise-bit.