Update README.md
Adding content to the "RefcardsProject" Repository on my @antiface GitHub account. I started the project roughly in 2012, and started this Repository a few years later, but never really wrote anything in it. I am now adding to it for the first time in a very long time. I talk about both the INNER and OUTER REFCARDS-PROJECT, similar to how there is an INNER and OUTER HISTORIOTHEQUE.