Update README.md
I added relevant content to this DAYPATHS folder in the "GeneralWorkflow" folder, in the METHODS folder, inside the "Documentation" Repository on my @antiface GitHub account. I started an UPDATE with today's date, to the README.md Markdown file, in this folder. Every folder in my GitHub repositories has a README.md Markdown file AS WELL AS my A.G. SIGNATURE, with hyperlink to my personal website, at alexgagnon.com.