Update README.md
Added new section/update on THE SWITCHBOARD METHOD FOR ART OPERATIONS using pseudocode to explain the "switching" process of the Chief Art Operator (a.k.a. "agent" or "machine/processor/worker") working on n number of jobs/processes/tasks. I may have used artificial intelligence to come up with some of the content of this folder in the "Documentation" GitHub repository by @antiface #historiotheque