Update README.md
Added a few things. Basically just added the word "PREAMBLE" at the top of the original portion of this document, plus in the first line about software projects and this "New Documentation", I added that it was specifically applied to contemporary, interdisciplinary art-research practices, which apparently was NOT the case when I originally came up with this project/experiment. But now it is DEFINITELY about contemporary art-research practices, but I'm keeping most of the original content/text I wrote in the first section.