Update README.md
I added a new section focusing more on the artist as "open collaborator", and how the artist was late to the "open source" practices game, due to the artist being previously seen mostly first as a genius who singlehandedly subverted the norms of society and rules of art, and then as an "entrepreneur" and "innovator". To protect copyright and trade secrets and whatnot, in an effort to maintain their competitive "edge", artists as independent contractors were late in adopting open source strategies of openness, transparency, and accountability.